+++ x/ecocredit/core/msg_update_class_admin.go
@@ -4,8 +4,6 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/x/auth/legacy/legacytx"
-
-	"github.com/regen-network/regen-ledger/x/ecocredit/v2"
 )
 
 var _ legacytx.LegacyMsg = &MsgUpdateClassAdmin{}
@@ -15,7 +13,7 @@ func (m MsgUpdateClassAdmin) Route() string { return sdk.MsgTypeURL(&m) }
 func (m MsgUpdateClassAdmin) Type() string { return sdk.MsgTypeURL(&m) }
 
 func (m MsgUpdateClassAdmin) GetSignBytes() []byte {
-	return sdk.MustSortJSON(ecocredit.ModuleCdc.MustMarshalJSON(&m))
+	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&m))
 }
 
 func (m *MsgUpdateClassAdmin) ValidateBasic() error {